Added a branch for Victor's work, and added code for min/max.
The min/max areas are stored as ranges of values for min/max.
eg. 'Max', or white pixel, is any value between say, 200 and 253.
These upper and lower values are calibrated based on the current
lighting conditions.
These ranges can be used to find white pixels in the form of 'if'
statements.