Added a branch for Victor's Work. Added min/max range functionality.
The min/max areas are stored as ranges of values.
eg. The 'max' area, or white pixel, is any pixel with a value between
227-253.
These values are calibrated and different based ont he current lighting
conditions for the track.
They may be used to find white pixels through the usage of 'if'
statements to check if the value is within a certain range.